Add Rides, Reservations, Positions and some facts
With this commit we introduce, Rides, Reservations (relations between
users and cars stored in the ReservationTable), and the Position class (with
its subclasses). Position class is abstract because we need some
generalization about start and end points for rides that must be either
SafeAreas or PowerGridStations.
Some facts add contraints on these classes.